fix: move estimate earlier
